% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feats.R
\name{add_feats}
\alias{add_feats}
\title{Add feats}
\usage{
add_feats(x, ...)
}
\arguments{
\item{...}{feat tables with names, i.e. genes=gene_df, snps=snp_df}
}
\description{
Add feats
}
\examples{
gggenomes \%>\%
  add_feats(genes=gene_df, snps=snp_df)
}
