% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_features}
\alias{add_features}
\title{Add features}
\usage{
add_features(x, ...)
}
\arguments{
\item{...}{feature tables with names, i.e. genes=gene_df, snps=snp_df}
}
\description{
Add features
}
\examples{
gggenomes \%>\%
  add_features(genes=gene_df, snps=snp_df)
}
