% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{add_seqs}
\alias{add_seqs}
\title{Add seqs}
\usage{
add_seqs(x, seqs, ...)
}
\arguments{
\item{x}{a gggenomes or gggenomes_layout objekt}

\item{seqs}{the sequences to add}

\item{...}{pass through to \code{as_seqs()}}
}
\value{
a gggenomes or gggenomes_layout object with added seqs
}
\description{
Add seqs
}
