% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfeats.R
\name{add_subfeats}
\alias{add_subfeats}
\title{Add subfeats}
\usage{
add_subfeats(x, parent_track_id, ..., transform = "none")
}
\arguments{
\item{x}{subfeat data convertible to a feat layout}

\item{parent_track_id}{track_id of the feats the subfeats map onto.}

\item{...}{subfeat tables with names, i.e. blast=blast_df, domains=domain_df}

\item{transform}{use if feats and subfeats are in different coordinate
spaces, i.e. if matching nucleotide-level annotations to protein level
annotations, e.g. genes and protein blast results.}

\item{.dots}{superceed dots with a list of arguments.}
}
\description{
Add subfeats
}
\examples{
gggenomes \%>\%
  add_subfeats(genes, blastp_hits, transform="aa2nuc")
}
