% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfeatures.R
\name{add_subfeatures}
\alias{add_subfeatures}
\title{Add subfeatures}
\usage{
add_subfeatures(x, parent_track_id, ...)
}
\arguments{
\item{...}{subfeature tables with names, i.e. blast=blast_df, domains=domain_df}
}
\description{
Add subfeatures
}
\examples{
gggenomes \%>\%
  add_features(genes=gene_df, snps=snp_df)
}
