% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sublinks.R
\name{add_sublinks}
\alias{add_sublinks}
\title{Add sublinks}
\usage{
add_sublinks(x, ..., .track_id = "genes", .transform = "aa2nuc")
}
\arguments{
\item{...}{sublink tables with names, i.e. blast=blast_df, domains=domain_df}

\item{.track_id}{track_id of the feats the sublinks map onto.}

\item{.transform}{one of "aa2nuc", "none", "nuc2aa"}
}
\description{
Add sublinks
}
