% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_contig.R
\name{as_contigs}
\alias{as_contigs}
\title{Coerce data into a tidy contig data frame}
\usage{
as_contigs(x, ...)
}
\arguments{
\item{x}{an object convertible to a `tbl_contig`}
}
\description{
A `tbl_contig` is a data frame for storing contig (or chromosome)
information for multiple genomes. Obligatory columns are `genome_id`,
`contig_id`, and contig `length`.
}
\examples{
chr <- tibble(
  genome_id = c(rep("A", 1), rep("B",2)),
  contig_id = c("a1", "b1", "b2"),
  length = c(5000,3000,1400))

as_contigs(chr)
}
