% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{as_factor_strand}
\alias{as_factor_strand}
\title{Convert numeric `1,-1,0` strand to `+,-,NA`-factor}
\usage{
as_factor_strand(strand)
}
\arguments{
\item{strand}{a vector encoding strandness}
}
\description{
Convert numeric `1,-1,0` strand to `+,-,NA`-factor
}
