% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_feature.R
\name{as_features}
\alias{as_features}
\title{Coerce data into feature tibble.}
\usage{
as_features(x, contigs, ..., everything = TRUE)
}
\arguments{
\item{x}{an object coercible to a `tbl_feature`}

\item{contigs}{a tbl_contig the features are associated with}

\item{...}{passed on to `layout()`}

\item{everything}{keep non-required columns}
}
\description{
Coerce data into feature tibble.
}
