% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genomes.R
\name{layout_genomes}
\alias{layout_genomes}
\title{Compute a genomes layout from sequences, features and links}
\usage{
layout_genomes(seqs = NULL, features = NULL, links = NULL,
  .feature_id = "genes", .link_id = "links", ...)
}
\arguments{
\item{seqs}{a table with sequence data (seq_id, bin_id, length)}

\item{features}{a table (or names list of tables) with feature data (seq_id,
bin_id, start, end)}

\item{links}{a table with link data (from, to, from_start, from_end,
to_start, to_end)}

\item{...}{layout parameters passed on to `layout_seqs()`}
}
\description{
See `gggenomes::gggenomes()` for more info.
}
