% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfeatures.R
\name{as_subfeatures}
\alias{as_subfeatures}
\title{Compute a layout for subfeature data}
\usage{
as_subfeatures(x, seqs, features, ..., everything = TRUE)
}
\arguments{
\item{x}{subfeature data convertible to a feature layout}

\item{seqs}{the sequence layout the parent features map onto.}

\item{features}{the parent features the subfeatures map onto.}

\item{...}{passed on to \code{layout_seqs()}}

\item{everything}{set to FALSE to drop optional columns}

\item{transform}{use if features and subfeatures are in different coordinate
spaces, i.e. if matching nucleotide-level annotations to protein level
annotations, e.g. genes and protein blast results.}
}
\value{
a tbl_df with plot coordinates
}
\description{
Read subfeature data such as domains or blast hits on genes into a tidy
dataframe. Subfeatures need to be associated with an already added feature
track. The subfeature track itself is internally converted into a new,
regular feature track by mapping the \code{start} and \code{end} coordinates provided
relative to their parent feature into coordinates relative to the sequences
underlying the parent features.
}
\details{
Obligatory columns are \code{feature_id}, \code{start} and \code{end}. Also recognized are
\code{strand} and \code{bin_id}.

Note \code{start} and \code{end} for every record will be coerced so that \code{start < end}. If no \code{strand} was provided, \code{strand} will be added and set to "+" for
records that initially had \code{start < end} and "-" for \code{end < start} inputs. If
\code{strand} was provided, \code{start} and \code{end} will be reorganized to conform with
\code{start < end} without any additional effect.
}
