% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genomes.R
\name{as_tbl_genomes.Granges}
\alias{as_tbl_genomes.Granges}
\title{Potential Future Work
Convert Granges objects}
\usage{
\method{as_tbl_genomes}{Granges}(x, ...)
}
\description{
Potential Future Work
Convert Granges objects
}
