% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{as_tracks}
\alias{as_tracks}
\title{Convert a list of tibbles into tracks with magic}
\usage{
as_tracks(tracks, tracks_exprs, reserved_ids = NULL)
}
\description{
Convert a list of tibbles into tracks with magic
}
\keyword{internal}
