% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{combine_strands}
\alias{combine_strands}
\title{Combine strands}
\usage{
combine_strands(strand, strand2, ...)
}
\arguments{
\item{strand}{first strand}

\item{strand2}{second strand}

\item{...}{more strands}
}
\value{
the combined strand
}
\description{
Combine strands
}
