% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{comma}
\alias{comma}
\title{Collapse a string with commas}
\usage{
comma(x, collapse = ",")
}
\arguments{
\item{x}{vector (coerced to character)}

\item{collapse}{character string to separate elements.}
}
\description{
Collapse a string with commas
}
\keyword{internal}
