% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{def_formats}
\alias{def_formats}
\title{Defined file formats and extensions}
\usage{
def_formats(
  file = NULL,
  ext = NULL,
  context = NULL,
  parser = NULL,
  allow_na = FALSE
)
}
\arguments{
\item{file}{a vector of file names}

\item{ext}{a vector of file extensions}

\item{context}{a vector of file contexts defined in
\code{gggenomes_global$def_formats}}

\item{parser}{a vector of file parsers defined in
\code{gggenomes_global$def_formats}}
}
\value{
dictionarish vector of file formats with recognized extensions as
names
}
\description{
For seamless reading of different file formats, gggenomes uses a mapping of
known formats to associated file extensions and contexts in which the
different formats can be read. The notion of context allows one to read
different information from the same format/extension. For example, a gbk file
holds both feature and sequence information. If read in "feats" context
\code{read_feats("*.gbk")} it will return a feature table, if read in "seqs"
context \code{read_seqs("*.gbk")}, a sequence index.
}
\examples{
# vector of defined zip formats and recognized extensions as names
# format of file
def_formats("foo.fa")

# formats associated with each extension
def_formats(ext=qc(fa, gff))

# all formats/extensions that can be read in seqs context; includes formats
# that are defined for context=NA, i.e. that can be read in any context.
def_formats(context="seqs")
}
