% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{def_types}
\alias{def_types}
\title{Default column types for defined formats}
\usage{
def_types(format)
}
\value{
a vector with default column types for the given format
}
\description{
Default column types for defined formats
}
