% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genome.R
\name{dim.tbl_genome}
\alias{dim.tbl_genome}
\title{`ggplot2::facet_null` checks data with `empty(df)` using `dim`. This causes
and error because dim(tbl_genome_layout) is undefined. Return dim of primary
table instead}
\usage{
\method{dim}{tbl_genome}(x)
}
\description{
`ggplot2::facet_null` checks data with `empty(df)` using `dim`. This causes
and error because dim(tbl_genome_layout) is undefined. Return dim of primary
table instead
}
