% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genomes.R
\name{dim.tbl_genomes}
\alias{dim.tbl_genomes}
\title{`ggplot2::facet_null` checks data with `emtpy(df)` using `dim`. This causes
and error because dim(tbl_genomes) is undefined. Return dim of primary table
instead}
\usage{
\method{dim}{tbl_genomes}(x)
}
\description{
`ggplot2::facet_null` checks data with `emtpy(df)` using `dim`. This causes
and error because dim(tbl_genomes) is undefined. Return dim of primary table
instead
}
