% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalebar.R
\name{draw_scalebar_grob}
\alias{draw_scalebar_grob}
\title{Draw a complete scalebar grob}
\usage{
draw_scalebar_grob(
  x0,
  x1,
  y,
  label = NULL,
  line,
  text,
  tick_height,
  label_offset,
  label_position = c("above", "below"),
  draw_ticks = TRUE,
  draw_label = TRUE
)
}
\description{
x0/x1/y are panel-relative coordinates in npc units.
}
\keyword{internal}
