% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feats.R
\name{drop_feat_layout}
\alias{drop_feat_layout}
\title{Drop feature layout}
\usage{
drop_feat_layout(x, keep = "strand")
}
\arguments{
\item{x}{feat_layout}

\item{keep}{features to keep}
}
\description{
Drop feature layout
}
