% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{drop_layout}
\alias{drop_layout}
\title{Drop a genome layout}
\usage{
drop_layout(data, ...)
}
\arguments{
\item{data}{layout}

\item{...}{additional data}
}
\value{
gggenomes object without layout
}
\description{
Drop a genome layout
}
