% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{drop_link_layout}
\alias{drop_link_layout}
\title{Drop a link layout}
\usage{
drop_link_layout(x, keep = "strand")
}
\arguments{
\item{x}{link_layout}

\item{keep}{features to keep}
}
\value{
link_layout without unwanted features
}
\description{
Drop a link layout
}
