% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{drop_seq_layout}
\alias{drop_seq_layout}
\title{Drop a seq layout}
\usage{
drop_seq_layout(x, keep = "strand")
}
\arguments{
\item{x}{seq_layout}

\item{keep}{features to keep}
}
\value{
seq_layout without unwanted features
}
\description{
Drop a seq layout
}
