% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emale_ava}
\alias{emale_ava}
\title{All-versus-all whole genome alignments of 6 EMALE genomes}
\format{
A data frame with 125 rows and 23 columns
\describe{
\item{file_id}{name of the file the data was read from}
\item{seq_id}{identifier of the sequence the feature appears on}
\item{length}{length of the sequence}
\item{start}{start of the feature on the sequence}
\item{end}{end of the feature on the sequence}
\item{strand}{orientation of the feature relative to the sequence (+ or -)}
\item{seq_id2}{identifier of the sequence the feature appears on}
\item{length2}{length of the sequence}
\item{start2}{start of the feature on the sequence}
\item{end2}{end of the feature on the sequence}
\item{map_match, map_length, map_quality, NM, ms, AS, nn, tp, cm, s1, de, rl, cg}{see \url{https://github.com/lh3/miniasm/blob/master/PAF.md} for additional columns}
}
}
\source{
\itemize{
\item Derived & bundled data: \code{ex("emales/emales.paf")}
}
}
\usage{
emale_ava
}
\description{
One row per alignment block. Alignments were computed with minimap2.
}
\keyword{datasets}
