% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emale_cogs}
\alias{emale_cogs}
\title{Clusters of orthologs of 6 EMALE proteomes}
\format{
A data frame with 48 rows and 3 columns
\describe{
\item{cluster_id}{identifier of the cluster}
\item{feat_id}{identifer of the gene}
\item{cluster_size}{number of features in the cluster}
}
}
\source{
\itemize{
\item Derived & bundled data: \code{ex("emales/emales-cogs.tsv")}
}
}
\usage{
emale_cogs
}
\description{
One row per feature. Clusters are based on manual curation.
}
\keyword{datasets}
