% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emale_gc}
\alias{emale_gc}
\title{Relative GC-content along 6 EMALE genomes}
\format{
A data frame with 2856 rows and 6 columns
\describe{
\item{file_id}{name of the file the data was read from}
\item{seq_id}{identifier of the sequence the feature appears on}
\item{start}{start of the feature on the sequence}
\item{end}{end of the feature on the sequence}
\item{name}{name of the feature}
\item{score}{relative GC-content of the window}
}
}
\source{
\itemize{
\item Derived & bundled data: \code{ex("emales/emales-gc.bed")}
}
}
\usage{
emale_gc
}
\description{
One row per 50 bp window.
}
\keyword{datasets}
