% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emale_prot_ava}
\alias{emale_prot_ava}
\title{All-versus-all alignments 6 EMALE proteomes}
\format{
A data frame with 827 rows and 13 columns
\describe{
\item{file_id}{name of the file the data was read from}
\item{feat_id}{identifier of the first feature in the alignment}
\item{feat_id2}{identifier of the second feature in the alignment}
\item{pident, length, mismatch, gapopen, start, end, start2, end2, evalue, bitscore}{see \url{https://github.com/seqan/lambda/wiki/BLAST-Output-Formats} for BLAST-tabular format columns}
}
}
\source{
\itemize{
\item Derived & bundled data: \code{ex("emales/emales-prot-ava.o6")}
}
}
\usage{
emale_prot_ava
}
\description{
One row per alignment. Alignments were computed with mmseqs2 (blast-like).
}
\keyword{datasets}
