% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emale_prot_uniref50}
\alias{emale_prot_uniref50}
\title{Alignments of 6 EMALE proteomes against Uniref50}
\format{
A data frame with 509 rows and 16 columns
\describe{
\item{file_id}{name of the file the data was read from}
\item{feat_id}{identifier of the first feature in the alignment}
\item{feat_id2}{identifier of the second feature in the alignment}
\item{pident, length, mismatch, gapopen, start, end, start2, end2, evalue, bitscore}{see \url{https://github.com/seqan/lambda/wiki/BLAST-Output-Formats} for BLAST-tabular format columns}
\item{seq_head}{full sequence header of the emale protein}
\item{seq_head2}{full sequence header of the Uniref50 protein}
\item{taxname}{one of the 4 most abundant taxonomic names among the Uniref50 hits or NA}
}
}
\source{
\itemize{
\item Derived & bundled data: \code{ex("emales/emales-prot-uniref50.tsv")}
}
}
\usage{
emale_prot_uniref50
}
\description{
One row per alignment. Alignments were computed with mmseqs2 (blast-like), and
filtered by evalue (<1e-20).
}
\keyword{datasets}
