% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emale_seqs}
\alias{emale_seqs}
\title{Sequence index of 6 EMALE genomes (endogenous virophages)}
\format{
A data frame with 6 rows and 4 columns
\describe{
\item{file_id}{name of the file the data was read from}
\item{seq_id}{sequence identifier}
\item{seq_desc}{sequence description}
\item{length}{length of the sequence}
}
}
\source{
\itemize{
\item Publication: \url{http://dx.doi.org/10.1101/2020.11.30.404863}
\item Raw data: \url{https://github.com/thackl/cb-emales}
\item Derived & bundled data: \code{ex("emales/emales.fna")}
}
}
\usage{
emale_seqs
}
\description{
A data set containing the sequence information on 6 endogenous virophages
found in the genomes of the marine protist \emph{Cafeteria burkhardae}.
}
\keyword{datasets}
