% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_files}
\alias{example_files}
\title{list files from example data sets}
\usage{
example_files(set = NULL, patt = "*", ...)
}
\arguments{
\item{set}{data set, see \code{example_sets} for a list}

\item{patt}{wildcard pattern to filter files}
}
\description{
list files from example data sets
}
