% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genome.R
\name{expose}
\alias{expose}
\title{Expose a single tibble of a tbl_genome object}
\usage{
expose(data, what)
}
\arguments{
\item{what}{data set to expose from tbl_genome_layout}
}
\description{
A tbl_genome is actually a list of tibbles. Expose lets you access a
particular tibble within the list.
}
