% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ext_to_format}
\alias{ext_to_format}
\title{File format from suffix}
\usage{
ext_to_format(x, context)
}
\arguments{
\item{x}{a vector of file extensions}

\item{context}{a file format context defined in \code{\link[=file_formats]{file_formats()}}}
}
\value{
a vector of formats with extensions as names
}
\description{
File format from suffix
}
\examples{
gggenomes:::ext_to_format(c("gff", "txt", "FASTA"), "feats")
}
