% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{file_exts}
\alias{file_exts}
\title{Defined file extensions and associated formats}
\usage{
file_exts(context)
}
\arguments{
\item{context}{a file format context defined in \code{gggenomes_global$file_formats}}
}
\value{
vector of file extensions with formats as names
}
\description{
Defined file extensions and associated formats
}
\examples{
# vector of zip-context file extensions and format names
gggenomes:::file_exts("zips")
}
