% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{file_formats}
\alias{file_formats}
\title{Defined file formats and recognized suffixes}
\usage{
file_formats(family)
}
\arguments{
\item{family}{a file format family defined in \code{gggenomes_global$file_formats}}
}
\value{
dictionarish vector of file formats with recognized suffixes as names
}
\description{
Defined file formats and recognized suffixes
}
\examples{
# vector of defined zip formats and recognized suffixes as names
file_formats("zips")
}
