% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{file_suffixes}
\alias{file_suffixes}
\title{Defined file suffixes and associated formats}
\usage{
file_suffixes(family)
}
\arguments{
\item{family}{a file format family defined in \code{gggenomes_global$file_formats}}
}
\value{
vector of file suffixes with formats as names
}
\description{
Defined file suffixes and associated formats
}
\examples{
# vector of zip-family file suffixes and format names
file_suffixes("zips")
}
