% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{flip_seq}
\alias{flip_seq}
\title{Flip sequences (contigs) by name or position}
\usage{
flip_seq(x, set, ...)
}
\arguments{
\item{x}{a gggenomes object}

\item{...}{One or more unquoted expressions separated by commas.
You can treat variable names like they are positions, so you can
use expressions like \code{x:y} to select ranges of variables.

Positive values select variables; negative values drop variables.
If the first expression is negative, \code{select()} will automatically
start with all variables.

Use named arguments, e.g. \code{new_name = old_name}, to rename selected variables.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in a context where column names
represent column positions. They also support
\link[rlang:quasiquotation]{unquoting} and splicing. See
\code{vignette("programming")} for an introduction to these concepts.

See \link[tidyselect:select_helpers]{select helpers} for more details and
examples about tidyselect helpers such as \code{starts_with()}, \code{everything()}, ...}
}
\description{
Invert the order and orientation of all specified sequences (contigs). Uses
dyplr::select syntax, which means unquoted names, positional arguments and
select helper functions, such as `starts_with()` are supported.
}
