% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{flip_strand}
\alias{flip_strand}
\title{Flip strand}
\usage{
flip_strand(strand, na = NA)
}
\arguments{
\item{strand}{some representation for strandedness}

\item{na}{what to use for \code{NA}}
}
\value{
the strand flipped
}
\description{
Flip strand
}
