% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_contig}
\alias{geom_contig}
\title{draw contigs}
\usage{
geom_contig(mapping = NULL, data = NULL, arrow = NULL, ...)
}
\arguments{
\item{data}{contig_layout}

\item{array}{set to non-NULL to generate default arrows}
}
\description{
draw contigs
}
