% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_feat.R
\name{geom_feat}
\alias{geom_feat}
\title{Draw feats}
\usage{
geom_feat(
  mapping = NULL,
  data = feats(),
  stat = "identity",
  position = "pile",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data}{feat_layout}
}
\description{
Draw feats
}
