% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_feature.R
\name{geom_feature}
\alias{geom_feature}
\title{Draw features}
\usage{
geom_feature(
  mapping = NULL,
  data = feats(),
  stat = "identity",
  position = "pile",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data}{feature_layout}
}
\description{
Draw features
}
