% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_gene_fast}
\alias{geom_gene_fast}
\title{Draw genes fast but not so pretty}
\usage{
geom_gene_fast(
  mapping = NULL,
  data = genes(),
  arrow = TRUE,
  nudge_by_strand = NULL,
  size = 4,
  color = "cornflowerblue",
  ...
)
}
\arguments{
\item{data}{feature_layout}
}
\description{
Draw genes fast but not so pretty
}
