% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gene_legacy.R
\name{geom_gene_legacy}
\alias{geom_gene_legacy}
\title{Draw gene models (legacy implementation)}
\usage{
geom_gene_legacy(
  mapping = NULL,
  data = genes(),
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  size = 2,
  rna_size = size,
  shape = size,
  rna_shape = shape,
  intron_shape = size,
  intron_types = c("CDS", "mRNA", "tRNA", "tmRNA", "ncRNA", "rRNA"),
  cds_aes = NULL,
  rna_aes = NULL,
  intron_aes = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{remove na values}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{size, rna_size}{the size of the gene model, aka the height of the
polygons. \code{rna_size} only applies to non-coding parts of the gene model,
defaults to size.}

\item{shape, rna_shape}{vector of height and width of the arrow tip, defaults
to size. If only one value is provided it is recycled. Set '0' to
deactivates arrow-shaped tips. \code{rna_shape} only applies to non-coding parts
of the gene model, defaults to shape.}

\item{intron_shape}{single value controlling the kink of the intron line.
Defaults to size. Set 0 for straight lines between exons.}

\item{intron_types}{introns will only be computed/drawn for features with
types listed here. Set to "CDS" to plot mRNAs as continuous features, and
set to NA to completely ignore introns.}

\item{cds_aes, rna_aes, intron_aes}{overwrite aesthetics for different model
parts. Need to be wrapped in \code{\link[ggplot2:aes]{ggplot2::aes()}}. NOTE: These remappings are
applied after the data has been transformed and mapped by the plot scales
(see \code{\link[ggplot2:aes_eval]{ggplot2::after_scale()}}). So you need to map between aesthetic names
(not data columns) and with standardized names, i.e. British English
spelling. These mappings can be used to dynamically change parts of the
gene model. For example, to change the color of introns from a hard-coded
"black" to the same color used to fill the CDS you could specify
\code{intron_aes=aes(colour = fill)}. By default, \code{rna_aes} is remapped with
\code{aes(fill=colorspace::lighten(fill, .5), colour=colorspace::lighten(colour, .5))} to give it a lighter appearance than the corresponding CDS but in the
same color.}

\item{...}{passed to layer params}
}
\value{
A ggplot2 layer with genes.
}
\description{
Legacy version of \code{\link[=geom_gene]{geom_gene()}} that draws one polygon per feature. It is
considerably slower for large numbers of genes, but kept as a fallback.
}
\keyword{internal}
