% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_link.R
\name{geom_link}
\alias{geom_link}
\title{Draw links}
\usage{
geom_link(
  mapping = NULL,
  data = links(),
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  offset = 0.1,
  ...
)
}
\arguments{
\item{offset}{distance between seq center and link start. Use two values
\verb{c(<offset_top>, <offset_bottom>)} for different top and bottom offsets}
}
\description{
Note \code{.adjacent_only=TRUE} is somewhat redundant with
layout_links(adjacent_only =TRUE), which currently is always set to TRUE
}
