% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_link_label}
\alias{geom_link_label}
\title{draw link labels}
\usage{
geom_link_label(
  mapping = NULL,
  data = links(),
  angle = 0,
  hjust = 0.5,
  vjust = 0.5,
  size = 4,
  repel = FALSE,
  ...
)
}
\description{
draw link labels
}
