% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gene_text.R
\docType{data}
\name{GeomFeatText}
\alias{GeomFeatText}
\title{Geom for feature text}
\format{
An object of class \code{GeomFeatText} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomFeatText
}
\description{
Geom for feature text
}
\keyword{datasets}
