% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny_to_link.R
\name{get_seqlen}
\alias{get_seqlen}
\title{Helper to extract seqlengths from DECIPHER::FindSyntney Object}
\usage{
get_seqlen(synteny_obj)
}
\arguments{
\item{synteny_obj}{A object produced by DECIPHER::FindSynteny}
}
\description{
Helper to extract seqlengths from DECIPHER::FindSyntney Object
}
\details{
Extracts the sequence lengths
}
