% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{get_seqs}
\alias{get_seqs}
\alias{set_seqs}
\title{Get/set the seqs track}
\usage{
get_seqs(x)

set_seqs(x, value)
}
\arguments{
\item{x}{a gggenomes or gggenomes_layout objekt}

\item{value}{to set vor seqs}
}
\value{
a gggenomes_layout track tibble
}
\description{
Get/set the seqs track
}
