% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{ggplot.gggenomes_layout}
\alias{ggplot.gggenomes_layout}
\title{ggplot.default tries to \code{fortify(data)} and we don't want that here}
\usage{
\method{ggplot}{gggenomes_layout}(data, mapping = aes(), ..., environment = parent.frame())
}
\description{
ggplot.default tries to \code{fortify(data)} and we don't want that here
}
