% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{ggplot.tbl_genome}
\alias{ggplot.tbl_genome}
\title{ggplot.default tries to `fortify(data)` and we don't want that}
\usage{
\method{ggplot}{tbl_genome}(data, mapping = aes(), ...,
  environment = parent.frame())
}
\description{
ggplot.default tries to `fortify(data)` and we don't want that
}
