% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{ggplot.tbl_genomes}
\alias{ggplot.tbl_genomes}
\title{ggplot.default tries to \code{fortify(data)}}
\usage{
\method{ggplot}{tbl_genomes}(data = NULL, mapping = aes(), ...,
  environment = parent.frame())
}
\description{
ggplot.default tries to \code{fortify(data)}
}
