% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{has_vars}
\alias{has_vars}
\title{Check if variables exist in object}
\usage{
has_vars(x, vars, any = FALSE)
}
\arguments{
\item{x}{object}

\item{vars}{variables to test}

\item{any}{if TRUE not all but at least one variable has to exists}
}
\value{
TRUE/FALSE
}
\description{
Returns TRUE if all variables exists. If `any=TRUE` returns TRUE if at least
one variable exists.
}
