% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{if_reverse}
\alias{if_reverse}
\title{Vectorised if_else based on strandedness}
\usage{
if_reverse(strand, reverse, forward)
}
\arguments{
\item{strand}{vector with strandedness information}

\item{reverse}{value to use for reverse elements}

\item{forward}{value to use for forward elements}
}
\description{
Vectorised if_else based on strandedness
}
