% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{is_reverse}
\alias{is_reverse}
\title{Check whether strand is reverse}
\usage{
is_reverse(strand, na = FALSE)
}
\arguments{
\item{strand}{some representation for strandedness}

\item{na}{what to use for \code{NA}}
}
\description{
Check whether strand is reverse
}
