% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout}
\alias{layout}
\title{Re-layout a genome layout}
\usage{
layout(x, ...)
}
\arguments{
\item{x}{layout}

\item{...}{additional data}
}
\value{
layout with updated scales
}
\description{
Re-layout the tracks and update the scales after seqs have been modified
}
