% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_contig.R
\name{layout.tbl_contig}
\alias{layout.tbl_contig}
\title{Layout tbl_contig}
\usage{
\method{layout}{tbl_contig}(x, ...)
}
\arguments{
\item{x}{feature data convertible to a feature layout}

\item{...}{not used}
}
\description{
Augment tbl_feature with all data necessary for plotting
}
