% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_feature.R
\name{layout.tbl_feature}
\alias{layout.tbl_feature}
\title{Layout features}
\usage{
\method{layout}{tbl_feature}(x, contigs, ...)
}
\arguments{
\item{x}{feature data convertible to a feature layout}

\item{...}{not used}
}
\description{
Augment features with all data necessary for plotting
}
