% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_link.R
\name{layout.tbl_link}
\alias{layout.tbl_link}
\title{Layout tbl_link}
\usage{
\method{layout}{tbl_link}(x, contigs, ...)
}
\arguments{
\item{x}{an object coercible to a `tbl_feature`}

\item{contigs}{a tbl_contig the features are associated with}

\item{...}{not used}
}
\description{
Layout tbl_link
}
