% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_contigs}
\alias{layout_contigs}
\title{layout contigs}
\usage{
layout_contigs(contigs, rubber_frac = 0.005)
}
\arguments{
\item{contigs}{table with genomic coordinates and required columns: gid, cid, clen, ...}

\item{rubber_frac}{relative amount of whitespace between contigs (based on largest genomes)}
}
\value{
contig_layout
}
\description{
layout contigs
}
