% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feats.R
\name{layout_feats}
\alias{layout_feats}
\title{Layout feats}
\usage{
layout_feats(
  x,
  seqs,
  keep = "strand",
  marginal = c("trim", "drop", "keep"),
  ...
)
}
\arguments{
\item{x}{feat data convertible to a feat layout}

\item{seqs}{the sequence layout the feat map onto.}

\item{marginal}{how to handle feats that stick out of sequences, for
example after focusing in on a subregion. Choices are to "drop" them, "keep"
them or "trim" them to the subregion boundaries.}

\item{...}{not used}
}
\description{
Augment feats with all data necessary for plotting
}
\keyword{internal}
