% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{layout_features}
\alias{layout_features}
\title{Layout features}
\usage{
layout_features(x, seqs, keep = "strand", marginal = c("trim", "drop",
  "keep"), ...)
}
\arguments{
\item{x}{feature data convertible to a feature layout}

\item{seqs}{the sequence layout the feature map onto.}

\item{marginal}{how to handle features that stick out of sequences, for
example after focusing in on a subregion. Choices are to "drop" them, "keep"
them or "trim" them to the subregion boundaries.}

\item{...}{not used}
}
\description{
Augment features with all data necessary for plotting
}
