% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{layout_seqs}
\alias{layout_seqs}
\title{Layout sequences}
\usage{
layout_seqs(
  x,
  spacing = 0.05,
  wrap = NULL,
  spacing_style = c("regular", "center", "spread"),
  keep = "strand"
)
}
\arguments{
\item{x}{seq_layout}

\item{spacing}{between sequences in bases (>1) or relative to longest bin (<1)}

\item{wrap}{wrap bins into multiple lines with at most this many nucleotides
per lin.}

\item{spacing_style}{one of "regular", "center", "spread"}

\item{keep}{keys to keep (default: "strand")}
}
\description{
Layout sequences
}
