% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{match_arg}
\alias{match_arg}
\title{match.arg defaulting to all choices}
\usage{
match_arg(arg, choices, several.ok = FALSE)
}
\description{
equivalent to base::match.arg, but returns all choices on arg=NULL if
several.ok=TRUE
}
\keyword{internal}
