% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalebar.R
\name{nice_scalebar_length}
\alias{nice_scalebar_length}
\title{Pick a nice scalebar length}
\usage{
nice_scalebar_length(x)
}
\arguments{
\item{x}{Positive numeric value.}
}
\value{
A numeric scalar.
}
\description{
Returns the largest 1, 2, or 5 x 10^n value <= x.
}
\keyword{internal}
