% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_stack2.R
\name{position_stack1}
\alias{position_stack1}
\alias{position_stack2}
\title{Stack overlapping features}
\usage{
position_stack1(offset = 0.02, strandwise = FALSE,
  vjust = as.numeric(strandwise), padding = 1, forward_below = FALSE,
  grouped = FALSE)

position_stack2(offset = 0.02, strandwise = TRUE,
  vjust = as.numeric(strandwise), padding = 1, forward_below = FALSE,
  grouped = FALSE)
}
\arguments{
\item{offset}{Shift overlapping labels up/down this much on the y-axis. The
y-axis distance between two sequences is 1, so this is usually a small
fraction, such as 0.02.}

\item{strandwise}{TRUE to stack forward features above and reverse features
below the sequence.}

\item{vjust}{How to align the stack relative to the sequence. 0 to center the
lowest stack level on the sequence, 1 to put forward/reverse sequence on half
offset above/below the sequence line.}

\item{padding}{A positive number to stack features that are not overlapping
but within this padding distance of each other, or negative number to not
stack features with smaller overlaps. Set to NA to not stack at all, but
still have strandwise or vjust take effect}

\item{forward_below}{By default forward features are stacked above, reverse
features depending on strandwise also above or below the genome. Set TRUE to
stack the other way around.}

\item{grouped}{stack features in same aestetics group as if they are one
feature. Useful for stacking multi-exon genes as a single unit}
}
\description{
`position_stack1()` stacks all overlapping features in one direction,
`position_stack2()` stacks forward and reverse strand features in opposite
directions above and below each sequence.
}
