% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_stack2.R
\name{position_stack2}
\alias{position_stack2}
\alias{position_stack3}
\title{Stack overlapping features}
\usage{
position_stack2(offset = 0.02, strandwise = FALSE,
  vjust = as.numeric(strandwise), padding = 1)

position_stack3(offset = 0.02, strandwise = TRUE,
  vjust = as.numeric(strandwise))
}
\description{
Stack overlapping features
}
