% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_alitv}
\alias{read_alitv}
\title{read an AliTV .json file
this usually contains sequences, links and (optionally) genes}
\usage{
read_alitv(file)
}
\value{
list
}
\description{
read an AliTV .json file
this usually contains sequences, links and (optionally) genes
}
