% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gffs_as_contigs}
\alias{read_gffs_as_contigs}
\title{Read genome ID, contig ID and contig lengths from .gff files. Parses
`##sequence-region` annotation (`rtracklayer` ignores those)}
\usage{
read_gffs_as_contigs(gff_files, gids = NULL)
}
\arguments{
\item{gids}{genome IDs to use with each file. If `NULL` infer from file name.}
}
\value{
A tibble with columns: 'gid', 'cid', 'length'.
}
\description{
Read genome ID, contig ID and contig lengths from .gff files. Parses
`##sequence-region` annotation (`rtracklayer` ignores those)
}
