% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_paf}
\alias{read_paf}
\title{read a .paf file (minimap/minimap2). Only the first 12 canonical
columns. Ignores tagged extra fields.}
\usage{
read_paf(paf_file)
}
\value{
tibble
}
\description{
read a .paf file (minimap/minimap2). Only the first 12 canonical
columns. Ignores tagged extra fields.
}
