% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{read_seqs}
\alias{read_seqs}
\alias{read_seq_len}
\alias{read_fai}
\title{Read a sequence index}
\usage{
read_seqs(files, format = NULL, .id = "file_id", ...)

read_seq_len(
  file,
  col_names = def_names("seq_len"),
  col_types = def_types("seq_len"),
  ...
)

read_fai(file, col_names = def_names("fai"), col_types = def_types("fai"), ...)
}
\arguments{
\item{file}{fasta or .fai/.seqkit.fai fasta index}

\item{col_names}{custom column names}
}
\value{
tibble
}
\description{
Read ID, description and length for each sequence from a fasta file or from a
samtools/seqkit fasta index file. Default columns are "seq_id", "seq_desc"
and "length".
}
\examples{
read_seqs("foo.fa")
read_seqs("foo.fa.fai")        # samtools style index, no description
read_seqs("foo.fa.seqkit.fai") # seqkit style with description
# custom col_names
read_seqs("foo.fa.fai", col_names=c("contig_id", "contig_desc", "len"))
}
