% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{require_vars}
\alias{require_vars}
\title{Require variables in an object}
\usage{
require_vars(x, vars, warn_only = FALSE)
}
\arguments{
\item{x}{object}

\item{vars}{required variables}

\item{warn_only}{don't die on missing vars}
}
\value{
the original tibble if all vars are present or warning only
}
\description{
Require variables in an object
}
