% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalebar.R
\name{resolve_scalebar_length}
\alias{resolve_scalebar_length}
\title{Resolve relative or absolute scalebar length}
\usage{
resolve_scalebar_length(length = 0.25, span)
}
\description{
\verb{0 < length < 1} is interpreted as a target fraction of \code{span}, rounded
down to a nice value. \code{length >= 1} is interpreted as absolute x units.
}
\keyword{internal}
