% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{scale_gggenomes}
\alias{scale_gggenomes}
\alias{scale_gggenomes_lab}
\alias{scale_gggenomes_numlab}
\title{y scale with genome labels}
\usage{
scale_gggenomes_lab(data, limits = NULL, ...)

scale_gggenomes_numlab(data, limits = NULL, ...)
}
\description{
y scale with genome labels
}
