% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_genomic.R
\name{scale_x_genomic}
\alias{scale_x_genomic}
\alias{format_bp}
\alias{label_bp}
\title{Genomic scales and base-pair labels}
\usage{
scale_x_genomic(
  ...,
  guide = "scalebar",
  unit = "bp",
  sep = " ",
  digits = 3,
  labels = NULL
)

format_bp(
  x,
  unit = "bp",
  sep = " ",
  digits = 3,
  prefixes = c(k = 1000, M = 1e+06, G = 1e+09),
  trim = TRUE,
  scientific = FALSE,
  ...
)

label_bp(unit = "bp", sep = " ", digits = 3)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}}

\item{unit}{unit suffix}

\item{sep}{separator between number and unit prefix+suffix}

\item{digits}{a positive integer indicating how many significant digits
    are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link[base]{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For more, notably the interpretation for complex numbers see
    \code{\link[base]{signif}}.)}

\item{x}{numeric base-pair value.}

\item{trim}{logical; if \code{FALSE}, logical, numeric and complex
    values are right-justified to a common width: if \code{TRUE} the
    leading blanks for justification are suppressed.}

\item{scientific}{either a logical specifying whether
    elements of a real or complex vector should be encoded in scientific
    format, or an integer penalty (see \code{\link[base]{options}("scipen")}).
    Missing values correspond to the current default penalty.}
}
\value{
A ggplot2 scale object with bp labels

Character scalar.

A labeller function for genomic data
}
\description{
Format genomic positions and scales using human-readable base-pair units.
\code{scale_x_genomic()} is the default genomic x-axis scale for gggenomes plots.
It wraps \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} with \code{\link[=guide_scalebar]{guide_scalebar()}} as its
default guide. See \code{\link[=axis_scalebar]{axis_scalebar()}} for convenient configuration and styling
of the scalebar. \code{format_bp()} formats numbers as base-pair quantities, and
powers \code{label_bp()}, the default \code{scale_x_genomic()} labelling function.
}
\section{Functions}{
\itemize{
\item \code{format_bp()}: format a number as human-readable base-pair value.

}}
\examples{
library(patchwork)

p0 <- gggenomes(genes = emale_genes) |> pick(1:2) + geom_gene() 
p1 <- p0 + scale_x_genomic("position in base-pairs", guide = "axis", unit = "", sep = "")
p2 <- p1 + scale_x_genomic(limits=c(-1000, 5000)) + theme_minimal()
p3 <- p0 + scale_x_continuous("regular x-axis")
p0 + p1 + p2 + p3 + plot_layout(ncol=2)
format_bp(c(0, 5e5, 1e6, 1.5e6, 2e6), digits=3, scientific=T)
}
