% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{select_set}
\alias{select_set}
\title{Select sets (genomes) by name or position}
\usage{
select_set(x, ...)
}
\arguments{
\item{x}{a gggenomes object}

\item{...}{One or more unquoted expressions separated by commas.
You can treat variable names like they are positions, so you can
use expressions like \code{x:y} to select ranges of variables.

Positive values select variables; negative values drop variables.
If the first expression is negative, \code{select()} will automatically
start with all variables.

Use named arguments, e.g. \code{new_name = old_name}, to rename selected variables.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in a context where column names
represent column positions. They also support
\link[rlang:quasiquotation]{unquoting} and splicing. See
\code{vignette("programming")} for an introduction to these concepts.

See \link[tidyselect:select_helpers]{select helpers} for more details and
examples about tidyselect helpers such as \code{starts_with()}, \code{everything()}, ...}
}
\description{
Choose which genomes to show and in which order. Uses dyplr::select syntax,
which means unquoted genome names, positional arguments and select helper
functions, such as `starts_with()` are supported. Renaming is not supported
at this point because changing the genome_id might break associations with
other tracks.
}
