% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{seqs}
\alias{seqs}
\title{Get/set the seqs track}
\usage{
seqs(x)
}
\arguments{
\item{x}{a gggenomes or gggenomes_layout objekt}
}
\value{
a gggenomes_layout track tibble
}
\description{
Get/set the seqs track
}
