% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\name{set_class}
\alias{set_class}
\alias{add_class}
\alias{strip_class}
\title{Modify object class attriutes}
\usage{
set_class(x, class, add = c("overwrite", "prepend", "append"))

add_class(x, class)

strip_class(x, class)
}
\arguments{
\item{x}{Object to assign new class to.}

\item{class}{Class value to add/strip.}

\item{add}{Possible values: "overwrite", "prepend", "append"}
}
\value{
Object x as class value.
}
\description{
Set class of an object. Optionally append or prepend to exiting class
attributes. \code{add_class} is short for \code{set_class(x, class, "prepend")}.
\code{strip_class} removes matching class strings from the class attribute vector.
}
