% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shift bins left/right}
\usage{
shift(x, bins = everything(), by = 0, center = FALSE)
}
\arguments{
\item{bins}{to shift left/right, select-like expression}

\item{by}{shift each bin by this many bases. Single value or vector of the
same length as bins.}
}
\description{
Shift bins along the x-axis, i.e. left or right in the default plot
layout. This is useful to align feats of interest in different bins.
}
\examples{
# Basic example plot
gggenomes(seqs = emale_seqs) |>
geom_seq() +
geom_bin_label() 

# All bins have been shifted 10000 basepaiers
gggenomes(seqs = emale_seqs) |>
shift(bins = everything(), by = 10000)
geom_seq() +
geom_bin_label() 

# Only RCC970_016B bin has been shifted for 5000 basepaiers
gggenomes(seqs = emale_seqs) |>
shift(bins = RCC970_016B, by = 5000, center = FALSE)
geom_seq() +
geom_bin_label() 
}
