% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{shift_set}
\alias{shift_set}
\title{Shift set (genomes) by a certain offset}
\usage{
shift_set(x, set, by)
}
\arguments{
\item{x}{a gggenomes object}

\item{set}{the set to shift}

\item{by}{shift by this much}
}
\description{
Shift set (genomes) by a certain offset
}
