% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{split_by}
\alias{split_by}
\title{Split by key preserving order}
\usage{
split_by(.data, key)
}
\arguments{
\item{key}{variable to split by}
}
\description{
Split by key column while preserving order according to the first
occurence. R base split converts keys to factors, changing default order to
alphanumeric.
}
\examples{
tibble(x=c(1,1,1,2), y=c("B", "A", "B", "B"), z="foo") \%>\%
  split_by(x)
}
\keyword{internal}
