% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{strand_chr}
\alias{strand_chr}
\title{Convert strand to character}
\usage{
strand_chr(strand, na = NA)
}
\arguments{
\item{strand}{some representation for strandedness}

\item{na}{what to use for \code{NA}}
}
\description{
Convert strand to character
}
