% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{strand_int}
\alias{strand_int}
\title{Convert strand to integer}
\usage{
strand_int(strand, na = NA)
}
\arguments{
\item{strand}{some representation for strandedness}

\item{na}{what to use for \code{NA}}
}
\description{
Convert strand to integer
}
