% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{strand_lgl}
\alias{strand_lgl}
\title{Convert strand to logical}
\usage{
strand_lgl(strand, na = NA)
}
\arguments{
\item{strand}{some representation for strandedness}

\item{na}{what to use for \code{NA}}
}
\description{
Convert strand to logical
}
