% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{suffix_to_format}
\alias{suffix_to_format}
\title{File format from suffix}
\usage{
suffix_to_format(x, family)
}
\arguments{
\item{x}{a vector of file suffixes}

\item{family}{a file format family defined in \code{\link[=file_formats]{file_formats()}}}
}
\value{
a vector of formats with suffixes as names
}
\description{
File format from suffix
}
\examples{
suffix_to_format(c("gff", "txt", "FASTA"), "feats")
}
