% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny_to_link.R
\name{synteny_to_link}
\alias{synteny_to_link}
\title{Format synteny-class into wide tibble This is for use with gggenomes}
\usage{
synteny_to_link(synteny_obj)
}
\arguments{
\item{synteny_obj}{A object produced by DECIPHER::FindSynteny}
}
\description{
Format synteny-class into wide tibble This is for use with gggenomes
}
\details{
This function takes a object of class synteny and converts it into a data.frame containing all synteny pairs
}
