% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genome_layout.R
\name{tbl_genome_layout}
\alias{tbl_genome_layout}
\alias{create_layout}
\alias{is.tbl_genome_layout}
\title{Genome Layout}
\usage{
tbl_genome_layout(chromosomes = NULL, features = list(), links = list(),
  ...)

create_layout(chromosomes = NULL, features = list(), links = list(), ...)

is.tbl_genome_layout(data)
}
\description{
Genome Layout
}
