% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_layout_contig.R
\name{tbl_layout_contig}
\alias{tbl_layout_contig}
\title{Layout contigs}
\usage{
tbl_layout_contig(contigs, rubber = 0.01, rubber_style = c("regular",
  "center", "spread"))
}
\arguments{
\item{rubber}{white space between contigs in bases, or relative to largest
genomes if <1.}

\item{rubber_style}{regular spread aligned}
}
\description{
Layout contigs
}
