% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_layout_link.R
\name{tbl_layout_link}
\alias{tbl_layout_link}
\title{layout links}
\usage{
tbl_layout_link(links, contig_layout)
}
\arguments{
\item{links}{links}

\item{contig_layout}{contig layout}
}
\value{
link_layout
}
\description{
layout links
}
