% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{theme_gggenomes_clean}
\alias{theme_gggenomes_clean}
\title{gggenomes default theme}
\usage{
theme_gggenomes_clean(
  base_size = 16,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
gggenomes default theme
}
