% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_ids}
\alias{track_ids}
\title{Named vector of track ids and types}
\usage{
track_ids(x, track_type, ...)
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}

\item{track_type}{restrict to any combination of "seqs", "feats" and "links".}

\item{...}{unused}
}
\value{
a named vector of track ids and types
}
\description{
Named vector of track ids and types
}
