% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_info}
\alias{track_info}
\title{Basic info on tracks in a gggenomes object}
\usage{
track_info(x, ...)
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}
}
\description{
Call on a gggenomes or gggenomes_layout object to get a short tibble
with ids, types, index and size of loaded tracks.
}
