% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_nrows}
\alias{track_nrows}
\title{Number of rows of all tracks tables}
\usage{
track_nrows(x, track_type = c("seqs", "features", "links"))
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}

\item{track_type}{restrict to any combination of "seqs", "features" and "links".}
}
\value{
a vector of number of rows all selected tracks tables
}
\description{
Number of rows of all tracks tables
}
