% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_type}
\alias{track_type}
\title{Track type by track id}
\usage{
track_type(x, track_id)
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}

\item{track_id}{id for which to get the track type}
}
\value{
a character string with the track type
}
\description{
Track type by track id
}
