% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_types}
\alias{track_types}
\title{All types of all tracks}
\usage{
track_types(x, types = c("seqs", "features", "links"))
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}

\item{types}{restrict to any combination of "seqs", "features" and "links".}
}
\value{
a vector of all types of all selected tracks
}
\description{
All types of all tracks
}
