% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{use}
\alias{use}
\title{Use a specific track table inside a `geom_*` call.}
\usage{
use(track = seqs, ...)
}
\arguments{
\item{...}{filter arguments passed through to [dplyr::filter].}
}
\description{
Use this function inside `geom_*` calls to set the track table of the
gggenomes layout you want to use, e.g. seqs, genes or links. Also allows you
to pass on filter arguments to subset the data.
}
