% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{use_genes}
\alias{use_genes}
\title{Use track as genes - magically augment with type/gene_id for geom_gene if
missing}
\usage{
use_genes(track_id = genes, ...)
}
\description{
Use track as genes - magically augment with type/gene_id for geom_gene if
missing
}
