% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tracks.R
\name{use_seqs}
\alias{use_seqs}
\alias{use_features}
\alias{use_links}
\alias{use_genes}
\alias{use_bins}
\title{Use a specific track table inside a `geom_*` call.}
\usage{
use_seqs(...)

use_features(.track_id = -1, ...)

use_links(.track_id = 1, ...)

use_genes(.track_id = 1, ...)

use_bins(..., group = vars())
}
\arguments{
\item{...}{Logical predicates passed on to [dplyr::filter].}

\item{.track_id}{The track to pull out, either as a literal variable name or
as a positive/negative integer giving the position from the left/right.}

\item{group}{what variables to use in grouping of bins from seqs in addition
to `y` and `bin_id`. Use this to get additional shared variables from the
seqs table into the bins table.}
}
\description{
Use these function inside `geom_*` calls to set the track table of the
gggenomes layout you want to use, e.g. seqs, features, links etc. Same
semantics as for `pull_*()` apply. `use_genes()` works on feature tracks, but
augments the data to play nicer with `geom_gene()`. `use_bins()` returns a
binwise summary of the seqs data, and powers `geom_bin_*()` calls. `...`
allow you to pass on filter arguments to subset the data.
}
