% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tracks.R
\name{vars_track}
\alias{vars_track}
\title{Tidyselect track variables}
\usage{
vars_track(
  x,
  track_id,
  track_type = c("seqs", "feats", "links"),
  ignore_pos = NULL
)
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}

\item{track_id}{a quoted or unquoted name or as positive/negative integer
giving the position from the left/right.}

\item{track_type}{restrict to these types of tracks - affects position-based
selection}

\item{ignore_pos}{names of tracks to ignore when selecting by position.}
}
\value{
The selected track_id as an unnamed string
}
\description{
Based on \code{tidyselect::vars_pull}. Powers track selection in \code{pull_track()}.
Catches and modifies errors from vars_pull to track-relevant info.
}
