% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{width}
\alias{width}
\alias{width0}
\title{The width of a range}
\usage{
width(start, end, base = 1)

width0(start, end, base = 0)
}
\arguments{
\item{start, end}{start and end of the range}

\item{base}{the base of the coordinate system, usually 1 or 0.}
}
\value{
a numeric vector
}
\description{
Always returns a positive value, even if start > end. `width0` is a short
handle for `width(..., base=0)`
}
