% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfeats.R
\name{add_subfeats}
\alias{add_subfeats}
\title{Add subfeats}
\usage{
add_subfeats(x, ..., .track_id = "genes", .transform = "aa2nuc")
}
\arguments{
\item{...}{subfeat tables with names, i.e. blast=blast_df, domains=domain_df}

\item{.track_id}{track_id of the feats the sublinks map onto.}

\item{.transform}{one of "aa2nuc", "none", "nuc2aa"}
}
\description{
Add subfeats
}
\examples{
genes <- tibble(seq_id="A", start=100, end=200, feat_id="gene1")
domains <- tibble(feat_id = "gene1", start=40, end=80)
gggenomes(genes=genes) \%>\% add_subfeats(domains) +
  geom_gene() + geom_feat()
}
