% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_feature.R
\name{as_features}
\alias{as_features}
\title{Coerce data into a feature tibble}
\usage{
as_features(x, seqs, ..., everything = TRUE)
}
\arguments{
\item{x}{an object coercible to a `tbl_feature`}

\item{...}{passed on to `layout()`}

\item{everything}{keep non-required columns}

\item{contigs}{a tbl_contig the features are associated with}
}
\description{
Coerce data into a feature tibble
}
