% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{as_features}
\alias{as_features}
\title{Compute a layout for feature data}
\usage{
as_features(x, seqs, ..., everything = TRUE)
}
\arguments{
\item{x}{feature data convertible to a feature layout}

\item{seqs}{the sequence layout the feature map onto.}

\item{...}{passed on to \code{layout_seqs()}}

\item{everything}{set to FALSE to drop optional columns}
}
\value{
a tbl_df with plot coordinates
}
\description{
Read feature data such as genes into a tidy dataframe and augment it with
layout information based on a sequence layout.
}
\details{
Obligatory columns are \code{seq_id}, \code{start} and \code{end}. Also recognized are
\code{strand} and \code{bin_id}.

Note \code{start} and \code{end} for every record will be coerced so that \code{start < end}. If no \code{strand} was provided, \code{strand} will added and set to "+" for
records that initially had \code{start < end} and "-" for \code{end < start} inputs. If
\code{strand} was provided, \code{start} and \code{end} will be ordered without any
additional effect.
}
