% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_feature.R
\name{as_features}
\alias{as_features}
\title{Compute a layout for feature data}
\usage{
as_features(x, seqs, ..., everything = TRUE)
}
\arguments{
\item{x}{feature data convertible to a feature layout}

\item{seqs}{the sequence layout the feature map onto.}

\item{...}{passed on to `layout_seqs()`}

\item{everything}{set to FALSE to drop optional columns}
}
\value{
a tbl_df with plot coordinates
}
\description{
Read feature data such as genes into a tidy dataframe and augment it with
layout information based on a sequence layout.
}
\details{
Obligatory columns are `seq_id`, `start` and `end`. Also recognized are
`strand` and `bin_id`.
}
