% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_link.R
\name{as_links}
\alias{as_links}
\title{Coerce data into a link tibble}
\usage{
as_links(x, contigs, ..., everything = TRUE)
}
\arguments{
\item{x}{an object coercible to a `tbl_feature`}

\item{contigs}{a tbl_contig the features are associated with}

\item{...}{passed on to `layout()`}

\item{everything}{keep non-required columns}
}
\description{
Coerce data into a link tibble
}
