% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{as_links}
\alias{as_links}
\title{Compute a layout for link data}
\usage{
as_links(x, seqs, ..., everything = TRUE)
}
\arguments{
\item{x}{link data convertible to a link layout}

\item{seqs}{the sequence layout the feat map onto.}

\item{...}{passed on to \code{layout_seqs()}}

\item{everything}{set to FALSE to drop optional columns}
}
\description{
Read link data of pairwise sequence or feat comparisons, such as
similarity searches into a tidy dataframe and augment it with layout
information based on a sequence layout.
}
\details{
Obligatory columns are \code{seq_id1} and \code{seq_id2}. Also recognized are
\code{start1}, \code{end1},\code{start2},\code{end2},\code{strand}, bin_id1\code{and}bin_id2.

During layouting, seq_id1,start1,end1 will be projected to x,xend,y, while
seq_id2,start2,end2 will be projected to xmin,xmax,yend. gggenomes uses these
maybe a bit odd names for the variables here, is so that they play nice with
ggplots native transformation functions for position aesthetics. Those only
work well with a specific set of predefined var names, which include those
used above.
}
