% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_feature.R
\name{as_numeric_strand}
\alias{as_numeric_strand}
\title{Encode strand numerically}
\usage{
as_numeric_strand(strand)
}
\arguments{
\item{strand}{a vector encoding strandness}
}
\description{
This functions converts common formats for strand encoding ("+/-/*",
"TRUE/FALSE/NA") into a `(1,-1,0)` representation. This numeric endoding
makes it easy to use strandness directly in arthmetic computations.
}
