% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subfeatures.R
\name{as_subfeatures}
\alias{as_subfeatures}
\title{Compute a layout for subfeature data}
\usage{
as_subfeatures(x, seqs, features, ..., everything = TRUE)
}
\arguments{
\item{x}{subfeature data convertible to a feature layout}

\item{seqs}{the sequence layout the parent features map onto.}

\item{features}{the parent features the subfeatures map onto.}

\item{...}{passed on to `layout_seqs()`}

\item{everything}{set to FALSE to drop optional columns}
}
\value{
a tbl_df with plot coordinates
}
\description{
Read subfeature data such as domains or blast hits on genes into a tidy
dataframe. Subfeatures need to be associated with an already added feature
track. The subfeature track itself is internally converted into a new,
regular feature track by mapping the `start` and `end` coordinates provided
relative to their parent feature into coordinates relative to the sequences
underlying the parent features .
}
\details{
Obligatory columns are `feature_id`, `start` and `end`. Also recognized are
`strand` and `bin_id`.

Note `start` and `end` for every record will be coerced so that `start <
end`. If no `strand` was provided, `strand` will be added and set to "+" for
records that initially had `start < end` and "-" for `end < start` inputs. If
`strand` was provided, `start` and `end` will be reorganized to conform with
`start < end` without any additional effect.
}
