% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand.R
\name{check_strand}
\alias{check_strand}
\title{Check strand}
\usage{
check_strand(strand, na)
}
\arguments{
\item{strand}{some representation for strandedness}

\item{na}{what to use for \code{NA}}
}
\value{
strand vector with unknown values replaced by \code{na}
}
\description{
Check strand
}
