% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{def_names}
\alias{def_names}
\title{Default column types for defined formats}
\usage{
def_names(format)
}
\value{
a vector with default column names for the given format
}
\description{
Default column types for defined formats
}
