% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{dim.gggenomes_layout}
\alias{dim.gggenomes_layout}
\title{`ggplot2::facet_null` checks data with `empty(df)` using `dim`. This causes
and error because dim(gggenome_layout) is undefined. Return dim of primary
table instead}
\usage{
\method{dim}{gggenomes_layout}(x)
}
\description{
`ggplot2::facet_null` checks data with `empty(df)` using `dim`. This causes
and error because dim(gggenome_layout) is undefined. Return dim of primary
table instead
}
