% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genomes.R
\name{expose}
\alias{expose}
\title{expose with non-standard evaluation of `what` argument.}
\usage{
expose(data, what)
}
\arguments{
\item{data}{tbl_genomes}

\item{what}{tbl to expose}
}
\description{
expose with non-standard evaluation of `what` argument.
}
