% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_feature}
\alias{geom_feature}
\title{draw features}
\usage{
geom_feature(
  mapping = NULL,
  data = use_features(),
  arrow = NULL,
  nudge_by_strand = NULL,
  size = 3,
  ...
)
}
\arguments{
\item{data}{feature_layout}
}
\description{
draw features
}
