% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gene.R
\name{geom_gene}
\alias{geom_gene}
\title{Draw genes}
\usage{
geom_gene(mapping = NULL, data = use_genes(1, type == "CDS"),
  stat = "identity", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, arrow_size = c(3, 3, 3),
  chevrons = NULL, ...)
}
\description{
Draw genes
}
