% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_gene}
\alias{geom_gene}
\title{draw genes}
\usage{
geom_gene(mapping = NULL, data = expose_data(features),
  nudge_by_strand = NULL, ...)
}
\arguments{
\item{mapping}{As
standard for ggplot2.}

\item{data}{feature_layout}

\item{...}{As
standard for ggplot2.}
}
\description{
draw genes
}
