% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_gene_label}
\alias{geom_gene_label}
\title{draw feature labels}
\usage{
geom_gene_label(mapping = NULL, data = use(genes), angle = 45,
  hjust = 0, nudge_y = 0.2, size = 2, ...)
}
\description{
draw feature labels
}
