% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_link}
\alias{geom_link}
\title{draw links}
\usage{
geom_link(mapping = NULL, data = NULL, nudge_frac = 0.1, ...)
}
\arguments{
\item{data}{link_layout}

\item{array}{set to non-NULL to generate default arrows}
}
\description{
draw links
}
