% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_seq}
\alias{geom_seq}
\title{draw seqs}
\usage{
geom_seq(mapping = NULL, data = use_seqs(), arrow = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{seq_layout}

\item{arrow}{set to non-NULL to generate default arrows}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
draw seqs
}
