% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_seq_label}
\alias{geom_seq_label}
\title{draw feat labels}
\usage{
geom_seq_label(
  mapping = NULL,
  data = seqs(),
  hjust = 0,
  nudge_y = -0.1,
  size = 6,
  ...
)
}
\description{
draw feat labels
}
