% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{gggenomes}
\alias{gggenomes}
\title{Plot genomes, features and synteny maps}
\usage{
gggenomes(seqs = NULL, features = NULL, links = NULL, ...,
  theme = c("clean", NULL), .layout = NULL)
}
\arguments{
\item{seqs}{a table with sequence data (seq_id, bin_id, length)}

\item{features}{a table (or names list of tables) with feature data (seq_id,
bin_id, start, end)}

\item{links}{a table with link data (from, to, from_start, from_end,
to_start, to_end)}

\item{...}{layout parameters passed on to `layout_genomes()`}

\item{theme}{choose a gggenomes default theme, NULL to omit.}

\item{.layout}{a pre-computed layout from `layout_genomes()` (advanced)}
}
\value{
gggenomes-flavored ggplot object
}
\description{
Sequence data: `read_fai`
}
\details{
Feature data: `read_gff`

Link data: `read_paf`
}
