% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R, R/seqs.R
\name{gggenomes}
\alias{gggenomes}
\alias{layout_genomes}
\alias{layout_seqs}
\title{Plot genomes, features and synteny maps}
\usage{
gggenomes(
  seqs = NULL,
  genes = NULL,
  features = NULL,
  links = NULL,
  ...,
  theme = c("clean", NULL),
  .layout = NULL
)

layout_genomes(
  seqs = NULL,
  genes = NULL,
  features = NULL,
  links = NULL,
  infer_bin_id = seq_id,
  infer_start = min(start, end),
  infer_end = max(start, end),
  infer_length = max(start, end),
  ...
)

layout_seqs(
  x,
  spacing = 0.05,
  wrap = NULL,
  spacing_style = c("regular", "center", "spread"),
  keep = "strand"
)
}
\arguments{
\item{seqs}{a table with sequence data (seq_id, bin_id, length)}

\item{genes}{a table or a list of table with gene data to be added as feature
tracks. Required columns: seq_id, bin_id, start, end.

For a single table, adds the track_id will be "genes". For a list, track_ids
are parsed from the list names, or if names are missing from the name of the
variable containing each table.}

\item{features}{same as genes, but the single table track_id will default to
"features".}

\item{links}{a table or a list of tables with link data to be added as link
tracks (columns: from, to, from_start, from_end, to_start, to_end). Same
naming scheme as for features.}

\item{...}{layout parameters passed on to [layout_genomes()] / [layout_seqs()]}

\item{theme}{choose a gggenomes default theme, NULL to omit.}

\item{.layout}{a pre-computed layout from [layout_genomes()]. Useful for
developmental purposes.}

\item{infer_length, infer_start, infer_end, infer_bin_id}{used to infer pseudo
seqs if only features or links are provided, or if no bin_id column was
provided. The expressions are evaluated in the context of the first feature
or link track.

By default subregions of sequences from the first to the last feature/link
are generated. Set `infer_start` to 0 to show all sequences from their
true beginning.}

\item{spacing}{between sequences in bases (>1) or relative to longest bin (<1)}

\item{wrap}{wrap bins into multiple lines with at most this many nucleotides
per lin.}
}
\value{
gggenomes-flavored ggplot object

gggenomes_layout object
}
\description{
Sequence data: `read_fai`
}
\details{
Feature data: `read_gff`

Link data: `read_paf`
}
\examples{
# Compare the genomic organization of three viral elements
# EMALEs: endogenous mavirus-like elements (example data shipped with gggenomes)
gggenomes(emale_seqs[c(1:2,6),], emale_genes, emale_tirs, emale_links) +
  geom_seq() + geom_bin_label() +                  # chromosomes and labels
  geom_feature(size=8) +                           # terminal inverted repeats
  geom_gene(aes(fill=strand), position="strand") + # genes
  geom_link(offset = 0.15)                         # synteny-blocks
}
