% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{gggenomes}
\alias{gggenomes}
\title{Plot genomes, features and synteny maps}
\usage{
gggenomes(contigs, genes = NULL, links = NULL, ..., scale = c("lab",
  "numlab", NULL), theme = c("clean", NULL), mapping = aes(),
  environment = parent.frame())
}
\arguments{
\item{...}{currently not used}

\item{scale}{choose a gggenomes dafault y-scale, NULL to omit.}

\item{theme}{choose a gggenomes default theme, NULL to omit.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{environment}{DEPRECATED. Used prior to tidy evaluation.}

\item{data}{genome layout generated by `layout_genomes()`}

\item{...}{layout parameter}
}
\value{
ggplot object
}
\description{
Plot genomes, features and synteny maps
}
