% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{gggenomes}
\alias{gggenomes}
\title{Plot genomes, features and synteny maps}
\usage{
gggenomes(
  genes = NULL,
  seqs = NULL,
  feats = NULL,
  links = NULL,
  .id = "file_id",
  spacing = 0.05,
  wrap = NULL,
  infer_bin_id = seq_id,
  infer_start = min(start, end),
  infer_end = max(start, end),
  infer_length = max(start, end),
  theme = c("clean", NULL),
  .layout = NULL,
  ...
)
}
\arguments{
\item{genes, feats}{A data.frame, a list of data.frames, or a character vector
with paths to files containing gene data. Each item is added as feature
track.

For a single data.frame the track_id will be "genes" and "feats",
respectively. For a list, track_ids are parsed from the list names, or if
names are missing from the name of the variable containing each data.frame.
Data columns:
\itemize{
\item required: \verb{seq_id,start,end}
\item recognized: \verb{strand,bin_id,feat_id,introns}
}}

\item{seqs}{A data.frame or a character vector with paths to files containing
sequence data. Data columns:
\itemize{
\item required: \verb{seq_id,length}
\item recognized: \verb{bin_id,start,end,strand}
}}

\item{links}{A data.frame or a character vector with paths to files
containing link data. Each item is added as links track. Data columns:
\itemize{
\item required: \verb{seq_id,seq_id2}
\item recognized: \verb{start,end,bin_id,start2,end2,bin_id2,strand}
}}

\item{.id}{The name of the column for file labels that are created when
reading directly from files. Defaults to "file_id". Set to "bin_id" if
every file represents a different bin.}

\item{spacing}{between sequences in bases (>1) or relative to longest bin (<1)}

\item{wrap}{wrap bins into multiple lines with at most this many nucleotides
per lin.}

\item{infer_length, infer_start, infer_end, infer_bin_id}{used to infer pseudo
seqs if only feats or links are provided, or if no bin_id column was
provided. The expressions are evaluated in the context of the first feat
or link track.

By default subregions of sequences from the first to the last feat/link
are generated. Set \code{infer_start} to 0 to show all sequences from their
true beginning.}

\item{theme}{choose a gggenomes default theme, NULL to omit.}

\item{.layout}{a pre-computed layout from \code{\link[=layout_genomes]{layout_genomes()}}. Useful for
developmental purposes.}
}
\value{
gggenomes-flavored ggplot object
}
\description{
Plot genomes, features and synteny maps
}
\examples{
# Compare the genomic organization of three viral elements
# EMALEs: endogenous mavirus-like elements (example data shipped with gggenomes)
gggenomes(emale_genes, emale_seqs, emale_tirs, emale_ava) +
  geom_seq() + geom_bin_label() +                  # chromosomes and labels
  geom_feat(size=8) +                              # terminal inverted repeats
  geom_gene(aes(fill=strand), position="strand") + # genes
  geom_link(offset = 0.15)                         # synteny-blocks

# with some more information
gggenomes(emale_genes, emale_seqs, emale_tirs, emale_ava) \%>\%
  add_feats(emale_ngaros, emale_gc) \%>\%
  add_clusters(emale_cogs) \%>\%
  flip_by_links() +
  geom_link(offset = 0.15, color="white") +                        # synteny-blocks
  geom_seq() + geom_bin_label() +                                  # chromosomes and labels
  # thistle4, salmon4, burlywood4
  geom_feat(size=6, position="identity") +                         # terminal inverted repeats
  geom_feat(data=feats(emale_ngaros), color="turquoise4", alpha=.3,
            position="strand", size=16) +
  geom_feat_note(aes(label=type), data=feats(emale_ngaros),
                 position="strand", nudge_y = .3) +
  geom_gene(aes(fill=cluster_id), position="strand") + # genes
  scale_fill_brewer("Conserved genes", palette="Dark2", na.value = "cornsilk3") +
  #scale_fill_viridis_b() +
  geom_ribbon(aes(x=(x+xend)/2, ymax=y+.24, ymin=y+.38-(.4*score),
                  group=seq_id, linetype="GC-content"), feats(emale_gc),
              fill="lavenderblush4", position=position_nudge(y=-.1))

# initialize plot directly from files
gggenomes(
  ex("emales/emales.gff"),
  ex("emales/emales.gff"),
  ex("emales/emales-tirs.gff"),
  ex("emales/emales.paf")
) + geom_seq() + geom_gene() + geom_feat() + geom_link()

# multi-contig genomes wrap to fixed width
s0 <- read_seqs(list.files(ex("cafeteria"), "Cr.*\\\\.fa", full.names = TRUE))
s1 <- s0 \%>\% filter(length > 5e5)
gggenomes(seqs=s1, infer_bin_id=file_id, wrap=5e6) +
  geom_seq() + geom_bin_label() + geom_seq_label()
}
