% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{gggenomes}
\alias{gggenomes}
\title{Plot genomes, features and synteny maps}
\usage{
gggenomes(seqs = NULL, features = NULL, links = NULL, ...,
  scale = c("lab", "numlab", NULL), theme = c("clean", NULL),
  mapping = aes(), environment = parent.frame())
}
\arguments{
\item{seqs}{a table with sequence data (seq_id, bin_id, length) or a layout
from `layout_genomes()`}

\item{...}{layout parameters passed on to `layout_genomes()`}

\item{scale}{choose a gggenomes dafault y-scale, NULL to omit.}

\item{theme}{choose a gggenomes default theme, NULL to omit.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{environment}{DEPRECATED. Used prior to tidy evaluation.}
}
\value{
gggenomes-flavored ggplot object
}
\description{
Sequence data: `read_fai`
}
\details{
Feature data: `read_gff`

Link data: `read_paf`
}
