% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{gggenomes}
\alias{gggenomes}
\title{Plot genomes, features and synteny maps}
\usage{
gggenomes(data = NULL, mapping = aes(), scale = c("lab", "numlab", NULL),
  scale_args = list(), theme = c("clean", NULL), theme_args = list(), ...,
  environment = parent.frame())
}
\arguments{
\item{data}{genome layout generated by `layout_genomes()`}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{scale}{choose a gggenomes dafault y-scale, NULL to omit.}

\item{theme}{choose a gggenomes default theme, NULL to omit.}

\item{...}{currently not used}

\item{environment}{If a variable defined in the aesthetic mapping is not
found in the data, ggplot will look for it in this environment. It defaults
to using the environment in which \code{ggplot()} is called.}
}
\value{
ggplot object
}
\description{
Plot genomes, features and synteny maps
}
