% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{gggenomes}
\alias{gggenomes}
\title{plotting genome feature and synteny maps}
\usage{
gggenomes(data = NULL, mapping = aes(), ...)
}
\arguments{
\item{data}{contig_layout generated by layout_genomes}

\item{...}{passed through to geom_contig}
}
\value{
ggplot object
}
\description{
plotting genome feature and synteny maps
}
