% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R, R/tbl_contig.R
\name{layout.tbl_contig}
\alias{layout.tbl_contig}
\title{Layout tbl_contig}
\usage{
\method{layout}{tbl_contig}(x, ...)

\method{layout}{tbl_contig}(x, ...)
}
\arguments{
\item{x}{an object coercible to a `tbl_feature`}

\item{...}{not used}

\item{...}{not used}
}
\description{
Augment tbl_feature with all data necessary for plotting

Augment tbl_feature with all data necessary for plotting
}
