% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_feature.R
\name{layout.tbl_feature}
\alias{layout.tbl_feature}
\title{Layout tbl_feature}
\usage{
\method{layout}{tbl_feature}(x, contigs, ...)
}
\arguments{
\item{x}{an object coercible to a `tbl_feature`}

\item{contigs}{a tbl_contig the features are associated with}

\item{...}{not used}
}
\description{
Augment tbl_feature with all data necessary for plotting
}
