% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{layout_features}
\alias{layout_features}
\title{Layout features}
\usage{
layout_features(x, seqs, keep = "feature_strand", ...)
}
\arguments{
\item{x}{feature data convertible to a feature layout}

\item{seqs}{the sequence layout the feature map onto.}

\item{...}{not used}
}
\description{
Augment features with all data necessary for plotting
}
