% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggenomes.R
\name{layout_genomes}
\alias{layout_genomes}
\title{Compute a genomes layout from sequences, features and links}
\usage{
layout_genomes(seqs = NULL, features = NULL, links = NULL,
  features_track_id = "genes", links_track_id = "links",
  infer_bin_id = seq_id, infer_start = min(start),
  infer_end = min(end), infer_length = max(end), ...)
}
\arguments{
\item{seqs}{a table with sequence data (seq_id, bin_id, length)}

\item{features}{a table (or names list of tables) with feature data (seq_id,
bin_id, start, end)}

\item{links}{a table with link data (from, to, from_start, from_end,
to_start, to_end)}

\item{features_track_id}{default ID for first features track}

\item{links_track_id}{default ID for first links track}

\item{infer_bin_id, infer_start, infer_end, infer_length}{used to infer pseudo
seqs if only features or links are provided. The expressions are evaluated in
the context of the first feature track, or the first links track.

By default subregions of sequences from the first to the last feature/link
are generated. Set `infer_start` to 0 to show all sequences from their
true beginning.}

\item{...}{layout parameters passed on to `layout_seqs()`}
}
\description{
See `gggenomes::gggenomes()` for more info.
}
