% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_links}
\alias{layout_links}
\title{layout links}
\usage{
layout_links(links, contig_layout)
}
\arguments{
\item{links}{links}

\item{contig_layout}{contig layout}
}
\value{
link_layout
}
\description{
layout links
}
