% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{layout_links}
\alias{layout_links}
\title{Layout tbl_link}
\usage{
layout_links(
  x,
  seqs,
  keep = "strand",
  adjacent_only = TRUE,
  marginal = c("trim", "drop", "keep"),
  ...
)
}
\arguments{
\item{x}{link data convertible to a link layout}

\item{seqs}{the sequence layout the feature map onto.}

\item{...}{not used}
}
\description{
Layout tbl_link
}
