% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{layout_seqs}
\alias{layout_seqs}
\title{Layout sequences}
\usage{
layout_seqs(x, spacing = 0.05, spacing_style = c("regular", "center",
  "spread"))
}
\arguments{
\item{x}{an object convertible to a sequence layout}

\item{spacing}{between sequences in bases (>1) or relative to longest bin (<1)}

\item{...}{not used}
}
\description{
Augment sequences with all data necessary for plotting
}
