% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_strandpile.R
\docType{data}
\name{position_strand}
\alias{position_strand}
\alias{position_pile}
\alias{position_strandpile}
\alias{position_sixframe}
\alias{PositionStrandpile}
\alias{PositionPile}
\alias{PositionStrand}
\alias{PositionSixframe}
\title{Stack features}
\usage{
position_strand(offset = 0.1, flip = FALSE, grouped = FALSE,
  base = offset/2)

position_pile(offset = 0.1, gap = 1, flip = FALSE, grouped = FALSE,
  base = offset)

position_strandpile(offset = 0.1, gap = 1, flip = FALSE,
  grouped = FALSE, base = offset * 1.5)

position_sixframe(offset = 0.1, flip = FALSE, grouped = FALSE,
  base = offset/2)
}
\arguments{
\item{offset}{Shift overlapping features up/down this much on the y-axis. The
y-axis distance between two sequences is 1, so this is usually a small
fraction, such as 0.1.}

\item{flip}{stack downward, and for stranded versions reverse upward.}

\item{grouped}{set TRUE to stack features in same aestetics group as if they
are one feature. Useful for stacking multi-exon genes as a single unit.}

\item{base}{How to align the stack relative to the sequence. 0 to center the
lowest stack level on the sequence, 1 to put forward/reverse sequence one
half offset above/below the sequence line.}

\item{gap}{If two features are closer together than this, they will be
stacked. Can be negative to allow small overlaps. NA disables stacking.}
}
\description{
`position_strand()` offsets forward features upward and reverse features
downward. `position_pile()` stacks overlapping features upward.
`position_strandpile()` stacks overlapping features up-/downward based on
their strand. `position_sixframe()` offsets the features based on their
strand and reading frame.
}
\keyword{datasets}
