% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{read_blast}
\alias{read_blast}
\title{Read BLAST tab-separated output}
\usage{
read_blast(
  file,
  col_names = def_names("blast"),
  col_types = def_types("blast"),
  comment = "#",
  swap_query = FALSE,
  ...
)
}
\arguments{
\item{col_names}{column names to use. Defaults to \link{def_names("blast")}
compatible with blast tabular output (\verb{--outfmt 6/7} in blast++ and \code{-m8}
in blast-legacy). \link{def_names("blast")} can easily be combined with extra
columns: \code{col_names = c(def_names("blast"), "more", "things")}.}

\item{swap_query}{if TRUE swap query and subject columns using \code{\link[=swap_query]{swap_query()}}
on import.}
}
\description{
Read BLAST tab-separated output
}
