% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feats.R
\name{read_feats}
\alias{read_feats}
\title{Read features from common file formats}
\usage{
read_feats(files, format = NULL, ...)
}
\value{
tibble
}
\description{
Read features from GFF3, Genbank, BED, BLAST tabular output or PAF files.
}
