% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_feats.R
\name{read_feats}
\alias{read_feats}
\title{Read features from common file formats}
\usage{
read_feats(files, format = NULL, .id = "file_id", ...)
}
\arguments{
\item{files}{files to reads. Should all be of same format.}

\item{format}{If NULL, will be guessed from file extension. Else, any format
known to gggenomes ("gff3", "gbk", ... see \code{\link[=file_formats]{file_formats()}} for full list)
or any suffix that maps to a known \verb{read_<suffix>} function to be called,
such as \code{\link[readr:read_tsv]{readr::read_tsv()}}, for example.}

\item{.id}{When binding output from several files, how to name the column
with the name of the file each record came from. Defaults to "file_id". Set
to "bin_id" if every file represents a different bin.}

\item{...}{additional arguments passed on to the specific read function for
the given format.}
}
\value{
a gggenomes compatible feature tibble
}
\description{
Read features from GFF3, Genbank, BED, BLAST tabular output or PAF files.
}
\examples{
# read a file
read_feats("data-raw/eden-utr.gff")
# read all gffs from a directory
read_feats(list.files("some/where", "*.gff$", full.names=TRUE))
# read remote files
gbk_phages <- c(
PSSP7 = "ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/858/745/GCF_000858745.1_ViralProj15134/GCF_000858745.1_ViralProj15134_genomic.gff.gz",
PSSP3 = "ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/904/555/GCF_000904555.1_ViralProj195517/GCF_000904555.1_ViralProj195517_genomic.gff.gz")
read_feats(gbk_phages)
}
