% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gff3.R
\name{read_gff3}
\alias{read_gff3}
\title{Read features from GFF3 files}
\usage{
read_gff3(file, sources = NULL, types = NULL)
}
\arguments{
\item{file}{toread}

\item{sources}{only return features from these sources}

\item{types}{only return features of these types, e.g. gene, CDS, ...}

\item{max_tags}{maximum number of optional fields to include}
}
\value{
tibble
}
\description{
Files with \verb{##FASTA} section work but result in parsing problems for all
lines of the fasta section. Just ignore those warnings, or strip the fasta
section ahead of time from the file.
}
