% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gffs_as_contigs}
\alias{read_gffs_as_contigs}
\title{Read genome_ids, contig_ids and contig lengths from .gff files.}
\usage{
read_gffs_as_contigs(gff_files, genome_ids = NULL)
}
\arguments{
\item{genome_ids}{to use with each file. If `NULL` infer from file name.}
}
\value{
A tibble with columns: genome_id, contig_id, length.
}
\description{
Parses `##sequence-region` annotation using `grep`. `rtracklayer` ignores
those lines.
}
