% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_paf}
\alias{read_paf}
\title{read a .paf file (minimap/minimap2). Only the first 12 canonical
columns. Ignores tagged extra fields.}
\usage{
read_paf(file, max_tags = 20)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \code{http://},
\code{https://}, \code{ftp://}, or \code{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[=clipboard]{clipboard()}} will read from the system clipboard.}

\item{max_tags}{maximum number of optional fields to include}
}
\value{
tibble
}
\description{
read a .paf file (minimap/minimap2). Only the first 12 canonical
columns. Ignores tagged extra fields.
}
