% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{read_seq_len}
\alias{read_seq_len}
\alias{read_fai}
\title{Read sequence index}
\usage{
read_seq_len(
  file,
  col_names = def_names("seq_len"),
  col_types = def_types("seq_len"),
  ...
)

read_fai(file, col_names = def_names("fai"), col_types = def_types("fai"), ...)
}
\description{
Read sequence index
}
\section{Functions}{
\itemize{
\item \code{read_seq_len}: read seqs from a single file in fasta, gbk or gff3 format.

\item \code{read_fai}: read seqs from a single file in seqkit/samtools fai format.
}}

