% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{read_seqs}
\alias{read_seqs}
\title{Read a sequence index}
\usage{
read_seqs(file, col_names = c("seq_id", "seq_desc", "length"))
}
\arguments{
\item{file}{fasta or .fai/.seqkit.fai fasta index}

\item{col_names}{custom column names}
}
\value{
tibble
}
\description{
Read ID, description and length for each sequence from a fasta file or from a
samtools/seqkit fasta index file. Default columns are "seq_id", "seq_desc"
and "length".
}
\examples{
read_seqs("foo.fa")
read_seqs("foo.fa.fai")        # samtools style index, no description
read_seqs("foo.fa.seqkit.fai") # seqkit style with description
# custom col_names
read_seqs("foo.fa.fai", col_names=c("contig_id", "contig_desc", "len"))
}
