% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{read_seqs}
\alias{read_seqs}
\alias{read_seq_len}
\alias{read_fai}
\title{Read a sequence index}
\usage{
read_seqs(files, format = NULL, .id = "file_id", parse_desc = TRUE, ...)

read_seq_len(
  file,
  col_names = def_names("seq_len"),
  col_types = def_types("seq_len"),
  ...
)

read_fai(file, col_names = def_names("fai"), col_types = def_types("fai"), ...)
}
\arguments{
\item{parse_desc}{remove \verb{key=some value} pairs from \code{seq_desc} and turn them
into \code{key}-named columns.}

\item{file}{fasta or .fai/.seqkit.fai fasta index}
}
\value{
A gggenomes-compatible sequence tibble
}
\description{
Read ID, description and length for each sequence from common formats
including FASTA, samtools/seqkit FASTA index files, and GFF3. Default columns
are seq_id, seq_desc and length.
}
\section{Functions}{
\itemize{
\item \code{read_seqs}: read seqs from files with automatic format detection

\item \code{read_seq_len}: read seqs from a single file in fasta, gbk or gff3 format.

\item \code{read_fai}: read seqs from a single file in seqkit/samtools fai format.
}}

\examples{
# read sequences from a fasta file.
read_seqs(ex("emales/emales.fna"), parse_desc=FALSE)

# read sequences from a fasta file with `parse_desc=TRUE` (default). `key=value`
# pairs are removed from `seq_desc` and parsed into columns with `key` as name
read_seqs(ex("emales/emales.fna"))

# from samtools/seqkit style index
read_seqs(ex("emales/emales.fna.seqkit.fai"))

# from multiple gff file
read_seqs(c(ex("emales/emales.gff"), ex("emales/emales-tirs.gff")))

}
