% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_seqs.R
\name{read_seqs}
\alias{read_seqs}
\alias{read_seq_len}
\alias{read_fai}
\title{Read a sequence index}
\usage{
read_seqs(files, format = NULL, .id = "file_id", ...)

read_seq_len(
  file,
  col_names = def_names("seq_len"),
  col_types = def_types("seq_len"),
  ...
)

read_fai(file, col_names = def_names("fai"), col_types = def_types("fai"), ...)
}
\arguments{
\item{file}{fasta or .fai/.seqkit.fai fasta index}
}
\value{
gggenomes-compatible seqs tibble
}
\description{
Read ID, description and length for each sequence from common formats
including FASTA, samtools/seqkit FASTA index files, and GFF3. Default columns
are "seq_id", "seq_desc" and "length".
}
\examples{
# from a fasta file
read_seqs("data-raw/emales/emales.fna")
# from samtools/seqkit style index
read_seqs("data-raw/emales/emales.fna.seqkit.fai")
# from a gff file
read_seqs("data-raw/emales/emales.gff")
}
