% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shift}
\alias{shift}
\title{Shift bins left/right}
\usage{
shift(x, bins = everything(), by = 0, center = FALSE)
}
\arguments{
\item{bins}{to shift left/right, select-like expression}

\item{by}{shift each bin by this many bases. Single value or vector of the
same length as bins.}
}
\description{
Shift bins along the x-axis, i.e. left or right in the default plot
layout. This is useful to align features of interest in different bins.
}
