% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{track_ids}
\alias{track_ids}
\title{Named vector of track ids and types}
\usage{
track_ids(x, ...)
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}

\item{types}{restrict to any combination of "seqs", "features" and "links".}
}
\description{
Named vector of track ids and types
}
