% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_genome.R
\name{use}
\alias{use}
\title{Use a (filtered) tibble inside a `geom_*` call.}
\usage{
use(what = contigs, ...)
}
\arguments{
\item{what}{data set to expose from tbl_genome_layout}

\item{...}{filter arguments passed through to [dplyr::filter].}
}
\description{
Use a (filtered) tibble inside a `geom_*` call.
}
