% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tracks.R
\name{vars_track}
\alias{vars_track}
\title{Tidyselect track variables}
\usage{
vars_track(x, track_id, types = c("seqs", "features", "links"))
}
\arguments{
\item{x}{A gggenomes or gggenomes_layout object}

\item{track_id}{as quoted or unquoted name or as positive/negative integer
giving the position from the left/right.}

\item{types}{restrict to these types of tracks - affects position-based
selection}
}
\value{
The selected track_id as an unnamed string
}
\description{
Based on `tidyselect::vars_pull`. Powers track selection in `pull_*/use*`.
Catches and modifies errors from vars_pull to track-relevant info.
}
