% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{x}
\alias{x}
\alias{xend}
\title{Project feature coordinates into layout space}
\usage{
x(start, end, strand, seq_x, seq_start, seq_strand)

xend(start, end, strand, seq_x, seq_start, seq_strand)
}
\arguments{
\item{seq_anchor}{the virtual x-start of the full length sequence in plot
space}
}
\description{
Project feature coordinates into layout space
}
